Conclusão do mapeamento

+++ QueryScoreBoard/QueryScoreBoard.Core/Entity/SQLMonitor/CheckinAnalysis.cs
@@ -4,9 +4,9 @@ namespace QueryScoreBoard.Core.Entity.SQLMonitor
 {
     public class CheckinAnalysis : EntityBase
     {
-        public decimal SqlId { get; set; }
-        public decimal PlanCost { get; set; }
-        public decimal Score { get; set; }
+        public int SqlId { get; set; }
+        public long PlanCost { get; set; }
+        public long Score { get; set; }
         public DateTime ExecutionDate { get; set; }
     }
 }
