+++ QueryScoreBoard/QueryScoreBoard.Core/Entity/SQLMonitor/ExecutionAnalysis.cs
@@ -2,10 +2,10 @@
 {
     public class ExecutionAnalysis : EntityBase
     {
-        public decimal SqlId { get; set; }
-        public decimal PlanCost { get; set; }
-        public decimal? ExecutionDuration { get; set; }
-        public decimal EnvironmentId { get; set; }
+        public int SqlId { get; set; }
+        public int PlanCost { get; set; }
+        public long? ExecutionDuration { get; set; }
+        public int EnvironmentId { get; set; }
         public string Comment { get; set; }
     }
 }
